Coalesce forks helper all in the same file
move all Aug 1st 2017 fork helpers in validation/forks.cpp
