+++ src/Makefile.am
@@ -193,7 +193,6 @@ BITCOIN_CORE_H = \
   txmempool.h \
   txorphanpool.h \
   ui_interface.h \
-  uahf_fork.h \
   undo.h \
   unlimited.h \
   stat.h \
@@ -291,7 +290,6 @@ libbitcoin_server_a_SOURCES = \
   txmempool.cpp \
   txorphanpool.cpp \
   tweak.cpp \
-  uahf_fork.cpp \
   unlimited.cpp \
   requestManager.cpp \
   validation/forks.cpp \
