--- src/uahf_fork.cpp
@@ -1,76 +1,0 @@
-// Copyright (c) 2017 The Bitcoin Unlimited Developers
-// Distributed under the MIT software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
-
-#include "uahf_fork.h"
-#include "chain.h"
-#include "chainparams.h"
-#include "primitives/block.h"
-#include "script/interpreter.h"
-#include "txmempool.h"
-#include "unlimited.h"
-
-#include <inttypes.h>
-#include <vector>
-
-bool IsTxProbablyNewSigHash(const CTransaction &tx)
-{
-    bool oldsighash = false;
-    for (auto txin : tx.vin)
-    {
-        std::vector<unsigned char> data;
-        CScript::const_iterator pc(txin.scriptSig.begin());
-        opcodetype op;
-        if (txin.scriptSig.GetOp(pc, op, data))
-        {
-            if (!data.empty())
-            {
-                if
-                    !(data.back() & SIGHASH_FORKID) { oldsighash = true; }
-            }
-        }
-    }
-    return (oldsighash == false);
-}
-
-bool IsTxUAHFOnly(const CTxMemPoolEntry &txentry)
-{
-    if ((txentry.sighashType & SIGHASH_FORKID) || (txentry.sighashType == 0))
-    {
-        return true;
-    }
-    return false;
-}
-
-// return true for every block from fork block and forward [consensusParams.uahfHeight,+inf)
-bool UAHFforkActivated(int height)
-{
-    const Consensus::Params &consensusParams = Params().GetConsensus();
-    if (height >= consensusParams.uahfHeight)
-    {
-        return true;
-    }
-    return false;
-}
-
-// This will check if the Fork will be enabled at the next block
-// i.e. we are at block x - 1, [consensusParams.uahfHeight-1, +inf]
-// state fork: enabled or activated
-bool IsUAHFforkActiveOnNextBlock(int height)
-{
-    const Consensus::Params &consensusParams = Params().GetConsensus();
-    if (height >= (consensusParams.uahfHeight - 1))
-        return true;
-    return false;
-}
-
-// return true only if 1st condition is true (Median past time > fork time)
-// and not the 2nd, i.e. we are at precisely [consensusParams.uahfHeight-1,consensusParams.uahfHeight-1]
-// state: fork enabled
-bool UAHFforkAtNextBlock(int height)
-{
-    const Consensus::Params &consensusParams = Params().GetConsensus();
-    if (height == (consensusParams.uahfHeight - 1))
-        return true;
-    return false;
-}
